Display nested Test Plan(s) in select drop-down on New Test Run page
also move discoverNestedTestPlans() into utils.js to minimize the
possibility of circular imports
